+++ src/ZaVoicePlayer.cpp
@@ -122,14 +122,14 @@ bool ZaPlayVoice(int voiceID, char *out_filename) {
 	const std::string& dir = g_zaConfig->ActiveGame->VoiceDir;
 	const std::string& preName = g_zaConfig->ActiveGame->VoiceName;
 	int index = 0;
-	for (; index < preName.size(); ++index) out_filename[index] = preName[index];
+	for (; index < (int)preName.size(); ++index) out_filename[index] = preName[index];
 
 	index += GetStrVoiceID(voiceID, out_filename + index);
 	out_filename[index++] = '.';
 
 	for (auto ext : g_zaConfig->ActiveGame->VoiceExt)
 	{
-		for (int i = 0; i < ext.size(); ++i) out_filename[index + i] = ext[i];
+		for (int i = 0; i < (int)ext.size(); ++i) out_filename[index + i] = ext[i];
 		out_filename[index + ext.size()] = 0;
 
 		std::string filePath = dir + '\\' + out_filename;
